rename src/bento/components/jwplayer -> src/bento/components/bento-jwplayer
