+++ src/bento/components/bento-jwplayer/1.0/component.js
